refactor: 重构模块导入路径和项目结构
- 统一模块导入路径为绝对路径，修复循环导入问题
- 重构项目结构，将主程序逻辑移动到main.py
- 优化配置文件加载错误处理，增加日志记录
- 重构服务初始化流程，确保单例模式正确实现
