--- VideoSlim/config.json
@@ -1,28 +1,0 @@
-{
-  "comment": "参数配置文件，不写会取用内置默认值。如果读取不到，说明配置文件不合法，请检查配置文件。具体参数的意思可以在README.md中看到",
-  "configs": {
-    "default": {
-      "x264": {
-        "crf": 23.5,
-        "preset": 8,
-        "I": 600,
-        "r": 4,
-        "b": 3
-      }
-    },
-    "custom_template": {
-      "x264": {
-        "crf": 30,
-        "preset": 8,
-        "I": 600,
-        "r": 4,
-        "b": 3
-      }
-    },
-    "default_gpu": {
-      "x264": {
-        "opencl_acceleration": true
-      }
-    }
-  }
-}
